refactor: 리뷰 반영 컴포넌트 및 타입, util 함수 수정
