+++ components/wine/WineMeta.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
-import vector from '@/assets/icons/vector.png';
+import arrowRight from '@/assets/icons/arrow-right.png';
 import WineRating from './WineRating';
-import { Wine } from '@/types/wine';
+import { Wine } from '@/types/domain/wine';
 
 type WineMetaProps = Pick<Wine, 'avgRating' | 'reviewCount'>;
 
@@ -9,7 +9,7 @@ export default function WineMeta({ avgRating, reviewCount }: WineMetaProps) {
   return (
     <div className="flex md:flex-col items-center md:items-end gap-9">
       <WineRating avgRating={avgRating} reviewCount={reviewCount} />
-      <Image src={vector} alt="arrow" width={24} height={24} />
+      <Image src={arrowRight} alt="arrow" width={24} height={24} />
     </div>
   );
 }
