+++ types/api/wine.dto.ts
@@ -1,0 +1,5 @@
+import { Wine } from '../domain/wine';
+
+export interface WineListResponse {
+  list: Wine[];
+}
