+++ utils/rating.ts
@@ -1,4 +1,7 @@
-export function getFilledStars(avgRating: number, max = 5) {
-  const filled = Math.round(avgRating);
-  return Array.from({ length: max }, (_, i) => i < filled);
+const MAX_STARS = 5;
+
+export function getFilledStars(avgRating: number) {
+  const safeRating = Number.isFinite(avgRating) ? avgRating : 0;
+  const filled = Math.min(MAX_STARS, Math.max(0, Math.round(safeRating)));
+  return Array.from({ length: MAX_STARS }, (_, i) => i < filled);
 }
